subscriber/requirements.txt
google-cloud-pubsub==2.17.1
redis==4.5.5
